feat: lib utils set some functions for server-only
