+++ package.json
@@ -18,6 +18,7 @@
     "qs": "^6.11.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "server-only": "^0.0.1",
     "sharp": "^0.32.5"
   },
   "devDependencies": {
